// src/interfaces/message-code.ts
export enum MessageCode {
  // User
  USER_CREATED = "USER_CREATED",
  USER_ID_REQUIRED = "USER_ID_REQUIRED",
  USER_ID_INVALID = "USER_ID_INVALID",
  USER_NOT_FOUND = "USER_NOT_FOUND",
  USER_FOUND = "USER_FOUND",
  USER_MAIL_REQUIRED = "USER_MAIL_REQUIRED",
  USER_MAIL_INVALID = "USER_MAIL_INVALID",
  USER_REMOVED = "USER_REMOVED",
  USER_UPDATED = "USER_UPDATED",
  USER_ALREADY_EXISTS = "USER_ALREADY_EXISTS",

  USER_FIRSTNAME_REQUIRED = "USER_FIRSTNAME_REQUIRED",
  USER_LASTNAME_REQUIRED = "USER_LASTNAME_REQUIRED",
  USER_PASSWORD_REQUIRED = "USER_PASSWORD_REQUIRED",

  // Category
  CATEGORY_CREATED = "CATEGORY_CREATED",
  CATEGORY_NAME_REQUIRED = "CATEGORY_NAME_REQUIRED",
  CATEGORY_DESCRIPTION_REQUIRED = "CATEGORY_DESCRIPTION_REQUIRED",
  CATEGORY_ALREADY_EXISTS = "CATEGORY_ALREADY_EXISTS",
  CATEGORIES_NOT_FOUND = "CATEGORIES_NOT_FOUND",
  CATEGORIES_FOUND = "CATEGORIES_FOUND",
  CATEGORY_FOUND = "CATEGORY_FOUND",
  CATEGORY_NOT_FOUND = "CATEGORY_NOT_FOUND",
  CATEGORY_ID_REQUIRED = "CATEGORY_ID_REQUIRED",
  CATEGORY_ID_INVALID = "CATEGORY_ID_INVALID",

  CATEGORY_REMOVED = "CATEGORY_REMOVED",
  CATEGORY_UPDATED = "CATEGORY_UPDATED",
}